requirements.txt

streamlit==1.31.0
google-generativeai==0.8.3
Pillow==10.2.0
streamlit-audio-recorder==0.0.8
pandas==2.2.0
plotly==5.18.0
